Fix #17080 Rename scaling IT test, avoid running at unit test
