+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-scaling/src/test/java/org/apache/shardingsphere/integration/data/pipline/cases/command/CreateTableSQLCommand.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.integration.data.pipline.cases.command.mysql;
+package org.apache.shardingsphere.integration.data.pipline.cases.command;
 
 import lombok.Data;
 
@@ -27,7 +27,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 @Data
 @XmlRootElement(name = "command")
 @XmlAccessorType(XmlAccessType.FIELD)
-public final class MySQLCommand {
+public final class CreateTableSQLCommand {
     
     @XmlElement(name = "create-table-order")
     private String createTableOrder;
